refactor: destination 도메인 분리
