+++ src/main/java/site/packit/packit/domain/destination/repository/DestinationRepository.java
@@ -1,8 +1,7 @@
-package site.packit.packit.domain.travel.repository;
+package site.packit.packit.domain.destination.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import site.packit.packit.domain.travel.entity.Destination;
-import site.packit.packit.domain.travel.entity.Travel;
+import site.packit.packit.domain.destination.entity.Destination;
 import site.packit.packit.global.exception.ResourceNotFoundException;
 
 import static site.packit.packit.domain.travel.exception.TravelErrorCode.DESTINATION_NOT_FOUND;
